+++ src/environments/environment.prod.ts
@@ -9,6 +9,14 @@ export const environment = {
     messagingSenderId: "1063364980313",
     appId: "1:1063364980313:web:3d19709c3673e18e638b48"
   },
+  firebaseKnowledgeBase: {
+    apiKey: "AIzaSyBeYnGKvE2wTERmB4-24BRP0chUMj0XCEM",
+    authDomain: "knowledgebase-store.firebaseapp.com",
+    projectId: "knowledgebase-store",
+    storageBucket: "knowledgebase-store.firebasestorage.app",
+    messagingSenderId: "342206336381",
+    appId: "1:342206336381:web:c1c6028ab57036fb761014"
+  },
   api: {
 
   }